Blog list spacing + renamed list component for clarity
